// src/app/dto/responseColaWebex.dto.ts
export class ResponseColaWebexDto {
  
  nIdSimVideCola: number;
  sNombres: string;
  sCodTicket: string;
  sOpeAsignado: string;
  nEstado: string;
  nIdSesion: string;
  bActivo:number;
  dFecCreado: string;
  dFecLlamada: string;
  nEdad:string;

  sLicenciaAsignar: string;

  nNombreOperador:string;
  dFecCreadoAMPM:string;
}